[ IN 23 ] rename files, add re-export for hooks, helpers
