+++ src/components/mail-verification-success/mail-verification-success.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image'
 import ConfirmImg from '@/assets/image/confirm-success.png'
 import { Button, Typography } from '@/components'
 import { PATH } from '@/consts/route-paths'
-import { useTranslation } from '@/hooks/use-translation'
+import { useTranslation } from '@/hooks'
 
 import s from './mail-verification-success.module.scss'
