+++ src/helpers/check-valid-query.ts
