+++ src/pages/auth/policy/index.tsx
@@ -1,11 +1,11 @@
 import { memo } from 'react'
 
-import s from './policy.module.scss'
-
 import { ArrowLeft } from '@/assets/icons'
 import { Button, Trans, Typography } from '@/components'
 import { PATH } from '@/consts/route-paths'
-import { useTranslation } from '@/hooks/use-translation'
+import { useTranslation } from '@/hooks'
+
+import s from './policy.module.scss'
 
 const Policy = memo(() => {
   const { t } = useTranslation()
