+++ src/schemas/index.ts
@@ -1,10 +1,10 @@
-export { forgotPasswordSchema } from './forgotPasswordSchema'
-export type { ForgotPasswordSchemaType } from './forgotPasswordSchema'
-export { loginSchema } from './loginSchema'
-export type { LoginFormValues } from './loginSchema'
-export { passwordRecoverySchema } from './passwordRecoverySchema'
-export type { PasswordRecoveryFormType } from './passwordRecoverySchema'
-export { profileSettingsSchema } from './profileSettingsSchema'
-export type { ProfileSettingsFormType } from './profileSettingsSchema'
-export { createRegisterSchema } from './registrationSchema'
-export type { RegisterFormType } from './registrationSchema'
+export { forgotPasswordSchema } from './forgot-password-schema'
+export type { ForgotPasswordSchemaType } from './forgot-password-schema'
+export { loginSchema } from './login-schema'
+export type { LoginFormValues } from './login-schema'
+export { passwordRecoverySchema } from './password-recovery-schema'
+export type { PasswordRecoveryFormType } from './password-recovery-schema'
+export { profileSettingsSchema } from './profile-setting-schema'
+export type { ProfileSettingsFormType } from './profile-setting-schema'
+export { createRegisterSchema } from './registration-schema'
+export type { RegisterFormType } from './registration-schema'
